api/requirements.txt
asgiref==3.4.1
backports.entry-points-selectable==1.1.1
certifi==2021.10.8
charset-normalizer==2.0.7
conda==4.3.16
distlib==0.3.3
Django==3.2.9
djangorestframework==3.12.4
filelock==3.3.2
idna==3.3
pipenv==2021.11.9
platformdirs==2.4.0
psycopg2-binary==2.9.2
pycosat==0.6.3
pytz==2021.3
requests==2.26.0
ruamel.yaml==0.17.17
ruamel.yaml.clib==0.2.6
sqlparse==0.4.2
urllib3==1.26.7
virtualenv==20.10.0
virtualenv-clone==0.5.7
